requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
qdrant-client==1.16.1
psycopg[binary]==3.1.18
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2